web: replace note creation date edit popover with dialog Signed-off-by: 01zulfi <85733202+01zulfi@users.noreply.github.com>
